Refactor into smaller and more self contained packages
Remove the dependencies from procfs and sysfs on xfs, nfs, bcache, and
disk iostats.  And make these package self contained so they
do not bring in any additional dependencies.

Also refactor the iostats package into a blockdevice package, and
improve naming and organization of this package.
