+++ sysfs/fs_test.go
@@ -28,85 +28,3 @@ func TestNewFS(t *testing.T) {
 		t.Error("want NewFS to fail if mount point is not a directory")
 	}
 }
-
-func TestFSXFSStats(t *testing.T) {
-	stats, err := FS(sysTestFixtures).XFSStats()
-	if err != nil {
-		t.Fatalf("failed to parse XFS stats: %v", err)
-	}
-
-	tests := []struct {
-		name      string
-		allocated uint32
-	}{
-		{
-			name:      "sda1",
-			allocated: 1,
-		},
-		{
-			name:      "sdb1",
-			allocated: 2,
-		},
-	}
-
-	const expect = 2
-
-	if l := len(stats); l != expect {
-		t.Fatalf("unexpected number of XFS stats: %d", l)
-	}
-	if l := len(tests); l != expect {
-		t.Fatalf("unexpected number of tests: %d", l)
-	}
-
-	for i, tt := range tests {
-		if want, got := tt.name, stats[i].Name; want != got {
-			t.Errorf("unexpected stats name:\nwant: %q\nhave: %q", want, got)
-		}
-
-		if want, got := tt.allocated, stats[i].ExtentAllocation.ExtentsAllocated; want != got {
-			t.Errorf("unexpected extents allocated:\nwant: %d\nhave: %d", want, got)
-		}
-	}
-}
-
-func TestFSBcacheStats(t *testing.T) {
-	stats, err := FS(sysTestFixtures).BcacheStats()
-	if err != nil {
-		t.Fatalf("failed to parse bcache stats: %v", err)
-	}
-
-	tests := []struct {
-		name   string
-		bdevs  int
-		caches int
-	}{
-		{
-			name:   "deaddd54-c735-46d5-868e-f331c5fd7c74",
-			bdevs:  1,
-			caches: 1,
-		},
-	}
-
-	const expect = 1
-
-	if l := len(stats); l != expect {
-		t.Fatalf("unexpected number of bcache stats: %d", l)
-	}
-	if l := len(tests); l != expect {
-		t.Fatalf("unexpected number of tests: %d", l)
-	}
-
-	for i, tt := range tests {
-		if want, got := tt.name, stats[i].Name; want != got {
-			t.Errorf("unexpected stats name:\nwant: %q\nhave: %q", want, got)
-		}
-
-		if want, got := tt.bdevs, len(stats[i].Bdevs); want != got {
-			t.Errorf("unexpected value allocated:\nwant: %d\nhave: %d", want, got)
-		}
-
-		if want, got := tt.caches, len(stats[i].Caches); want != got {
-			t.Errorf("unexpected value allocated:\nwant: %d\nhave: %d", want, got)
-		}
-	}
-}
