+++ xfs/parse_test.go
@@ -18,7 +18,6 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/prometheus/procfs"
 	"github.com/prometheus/procfs/xfs"
 )
 
@@ -425,7 +424,7 @@ func TestParseStats(t *testing.T) {
 			stats, err = xfs.ParseStats(strings.NewReader(tt.s))
 		}
 		if tt.fs {
-			stats, err = procfs.FS("../fixtures/proc").XFSStats()
+			stats, err = xfs.ReadProcStat("../fixtures/proc")
 		}
 
 		if tt.invalid && err == nil {
